Migrate `#dom/radiobuttons` to typescript; documentation
